requirements.txt
aiohttp==3.6.2
altgraph==0.17       
astroid==2.4.2       
async-timeout==3.0.1 
attrs==19.3.0        
beautifulsoup4==4.9.1
certifi==2020.6.20   
cffi==1.14.0
chardet==3.0.4       
click==7.1.2
colorama==0.4.3      
discord.py==1.3.4    
dnspython==2.0.0     
dominate==2.5.1      
email-validator==1.1.1
flake8==3.8.3
flash==1.0.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
pefile==2019.4.18
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.5.3
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyNaCl==1.3.0
PySocks==1.7.1
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.9
visitor==0.1.3
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
yarl==1.4.2